refactor(app,ui): rework api-keys settings — header Create, compact row menu, drop security notice
