+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/api-keys/_components/org-api-key-list-loading.tsx
@@ -1,26 +1,26 @@
 import { Skeleton } from "@repo/ui/components/ui/skeleton";
 
+const ROWS = [
+  { name: "w-40", detail: "w-64" },
+  { name: "w-32", detail: "w-56" },
+  { name: "w-36", detail: "w-60" },
+];
+
 export function OrgApiKeyListLoading() {
   return (
-    <div className="space-y-4">
-      <div className="flex items-center justify-between">
-        <Skeleton className="h-5 w-32" />
-        <Skeleton className="h-9 w-32" />
-      </div>
-      <div className="overflow-hidden rounded-lg border border-border/60">
-        {Array.from({ length: 3 }).map((_, i) => (
-          <div
-            className="flex items-center justify-between border-border/60 border-b px-4 py-4 last:border-b-0"
-            key={i}
-          >
-            <div className="space-y-2">
-              <Skeleton className="h-4 w-40" />
-              <Skeleton className="h-3 w-56" />
-            </div>
-            <Skeleton className="h-8 w-8" />
+    <div className="divide-y divide-border/60 overflow-hidden rounded-lg border border-border/60">
+      {ROWS.map((row, i) => (
+        <div
+          className="flex items-center justify-between gap-4 px-4 py-4"
+          key={i}
+        >
+          <div className="space-y-2">
+            <Skeleton className={`h-4 ${row.name}`} />
+            <Skeleton className={`h-3 ${row.detail}`} />
           </div>
-        ))}
-      </div>
+          <Skeleton className="h-6 w-6 rounded-md" />
+        </div>
+      ))}
     </div>
   );
 }
