+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/api-keys/page.tsx
@@ -1,8 +1,8 @@
 import { HydrateClient, prefetch, trpc } from "@repo/app-trpc/server";
 import { Suspense } from "react";
+import { OrgApiKeyCreate } from "./_components/org-api-key-create";
 import { OrgApiKeyList } from "./_components/org-api-key-list";
 import { OrgApiKeyListLoading } from "./_components/org-api-key-list-loading";
-import { SecurityNotice } from "./_components/security-notice";
 
 export const dynamic = "force-dynamic";
 
@@ -11,14 +11,17 @@ export default function OrgApiKeysPage() {
 
   return (
     <div className="space-y-8">
-      <div>
-        <h2 className="font-medium font-pp text-2xl text-foreground">
-          API Keys
-        </h2>
-        <p className="mt-1 text-muted-foreground text-sm">
-          Manage API keys for programmatic access to your organization's
-          resources.
-        </p>
+      <div className="flex items-start justify-between gap-4">
+        <div>
+          <h2 className="font-medium font-pp text-2xl text-foreground">
+            API Keys
+          </h2>
+          <p className="mt-1 text-muted-foreground text-sm">
+            Manage API keys for programmatic access to your organization's
+            resources.
+          </p>
+        </div>
+        <OrgApiKeyCreate />
       </div>
 
       <HydrateClient>
@@ -26,8 +29,6 @@ export default function OrgApiKeysPage() {
           <OrgApiKeyList />
         </Suspense>
       </HydrateClient>
-
-      <SecurityNotice />
     </div>
   );
 }
