+++ packages/ui/src/components/ui/button.tsx
@@ -31,6 +31,7 @@ const buttonVariants = cva(
         lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
         xl: "h-12 rounded-md px-8 has-[>svg]:px-6",
         icon: "size-9",
+        "icon-sm": "size-6",
       },
     },
     defaultVariants: {
